:sparkles: Root화면에서 이동전환 로직 가구현
