+++ MashUp-iOS/MashUp-iOS/Source/Screen/Root/RootController.swift
@@ -33,15 +33,46 @@ final class RootController: BaseViewController, ReactorKit.View {
     
     private func consume(_ reactor: Reactor) {
         reactor.pulse(\.$step).compactMap { $0 }
-        .subscribe(onNext: { step in
+        .subscribe(onNext: { [weak self] step in
             switch step {
             case .splash:
-                ()
+                self?.presentSplashViewController()
+                
             case .signIn:
-                ()
+                self?.switchToSignInViewController()
+                
             case .home(let userSession):
-                ()
+                self?.switchToHomeViewController(with: userSession)
             }
         }).disposed(by: self.disposeBag)
     }
+    
+    private func presentSplashViewController() {
+        let viewController = SplashViewController()
+        self.present(viewController, animated: false, completion: nil)
+    }
+    
+    private func switchToSignInViewController() {
+        let signInViewController = SignInViewController()
+        if let presentedViewController = self.presentedViewController {
+            presentedViewController.dismiss(animated: false, completion: {
+                self.present(signInViewController, animated: false, completion: nil)
+            })
+        } else {
+            self.present(signInViewController, animated: false, completion: nil)
+        }
+    }
+    
+    private func switchToHomeViewController(with userSession: UserSession) {
+        let homeViewContorller = HomeViewContorller()
+        if let presentedViewController = self.presentedViewController {
+            presentedViewController.dismiss(animated: false, completion: {
+                self.present(homeViewContorller, animated: false, completion: nil)
+            })
+        } else {
+            self.present(homeViewContorller, animated: false, completion: nil)
+        }
+    }
 }
+
+class HomeViewContorller: UIViewController {}
